// apps/web/src/lib/backend/generated/EntryData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "../serde_json/JsonValue";

export type EntryData = { 
/**
 * Path to the entry.
 */
path: string, 
/**
 * Title from frontmatter.
 */
title: string | null, 
/**
 * All frontmatter properties.
 */
frontmatter: { [key in string]?: JsonValue }, 
/**
 * Body content (after frontmatter).
 */
content: string, };